Refactor: move global.config.checker to issue.config
- CircularDependencyChecker에 대한 config를 issue.config 아래로 이동
