--- backend/src/main/java/com/tissue/api/global/config/checker/CheckerType.java
@@ -1,6 +1,0 @@
-package com.tissue.api.global.config.checker;
-
-public enum CheckerType {
-	DFS,
-	CACHED_DFS
-}
