+++ backend/src/main/java/com/tissue/api/issue/config/CircularDependencyCheckerType.java
@@ -1,0 +1,6 @@
+package com.tissue.api.issue.config;
+
+public enum CircularDependencyCheckerType {
+	DFS,
+	CACHED_DFS
+}
